Bug 559478 - [tests] Delete obsolete ShellClosingTest
The existing ShellClosingTest depends on test execution order i.e. it
requires another test to open and not cleanup a quick access dialog.

Due to the heavy changes on quick access the test can be removed. It
covered a problem with the old quick access search field which is not
used anymore.

Change-Id: I573e588e45e3cdb85e61f515988c5f01fc5c4975
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>