Bug 541447 - [Tests] testCancelWizardCancelsJob fails because of unexpected filesystem root
Test used File.listRoots()[0] to get a root directory which will likely
take a long time to scan. While this will probably always work on Unix
like systems and most of the time on Windows.
But it can happen that the first root on Windows isn't C:\ as most would
expect but A:\ or B:\ which are traditional used for floppy drives and
permitted to be used for any removable drive and in fact A:\ is returned
on integration build's Windows test machine as first root.

Change-Id: Ic368276a19047fa81ba7d3e3e756783b2eb597ed
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>