Bug 559758 - Only present unique launch configurations
Try to identify unique launch configurations based on kind and name and
return the unique set of configurations.
To avoid comparing every launch configuration with all others, map the
name of the configuration to the configuration itself and then as a
second step if there are duplicated names, do a another pass and verify
that there are only one configuration with the same attribute set (both
keys and values are compared) in the resulting list of configurations.

Change-Id: Id72d2842b096332fa73b3cc23e1c94531f6df28e
Signed-off-by: Torbjörn Svensson <azoff@svenskalinuxforeningen.se>