Use map method directly if possible
JDT provides a new cleanup which allows to transform code to call
methods directly on map instead of accessing the values or the keys
before calling the same method.

Saves code which we have to read and might be slightly faster.

Includes also some manual adjustments in TextEditBasedChangeManager and
RefactoringSearchEngine2 based on the review feedback.

Change-Id: I22632bc1dd6a6ff1bd687c93d5f3f688bee718b2
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>