Bug 553483 - ViewerDropAdapter does not reset lastValidOperation
In clearState() the lastValidOperation and overrideOperation is reset.
Furthermore in dragEnter() clearState() is called on entry (this could
also be moved to client code, see comment code).
The faulty scenario to consider is: a drag source, that does only
support DROP_COPY and a drop target, that accepts DROP_MOVE and
DROP_COPY. If then the first D&D was a DROP_COPY, the second one (from
the same 'copy-only' source) would be treated (and accepted) as
DROP_COPY, even if it was DROP_NONE, because the DROP_COPY is
(non-erasable) stored in lastValidOperation (=> surprising behavior for
clients...).

Change-Id: Ib0c491585fcca9a1871491caedb05f5aa97aa093
Signed-off-by: Gerhard Kreuzer <gerhard.kreuzer@siemens.com>