Bug 363009 - Refactor Infocenter deployment as webapp
This change refactors the approach to deploy the Eclipse information
center as web application on web application servers like Tomcat.

The previous approach involved several error prone manual steps. To
create a consistent OSGi application within the web archive the
application is now defined as Eclipse product and built with Tycho.

The web application is automatically assembled from the application
which makes it obsolete to manually search for a fitting version of the
servletbridge bundles and copying them to the application. All is done
within the application build.

This change refactors the project structure from root directory
'web-archive' to 'infocenter-web'.

A README.md file is provided with instructions for building the
infocenter application with the new approach.

Change-Id: I52ae067fd7eeccf669243942134ebac3904a3bb4
Signed-off-by: Karsten Thoms <karsten.thoms@karakun.com>