Bug 550898 - Allow to create subfolder in the New File wizard
Added the code for validation in the creation of a new file. If the
users enter "file.txt", the input field is validated and (which was the
previous behaviour).
If the users enter "path/to/file.txt" then the entire path is checked
for validation, and the file is created if it has a valid file name.

Change-Id: I2255125348bbd383f7f603626855944e4ca2c291
Signed-off-by: ejallow <jallowebrima7@gmail.com>