Bug 426557 - [KeyBindings] Welcome Screen disables shortcuts
Keybindings were bound to the limbo shell for parts that are invisible
when the are constructed, e.g. when it is contained in an invisible
container such as a minimized area.

As of commit cb265e3b935816c99167c2ec7864f982ad8caf3e the
LegacyHandlerService for a PartSite is bound to the active part only.
Therefore, the KeyBindingService no longer needs to add another check
for the active Shell and Site. By removing this expression, keybindings
are no longer bound to the limbo shell.

Furthermore, the default expression was inserted twice in the
HanderService when no expression was provided.

Change-Id: I31499b41d0c9df8f18bc210971f16ab5468fc81c
Signed-off-by: Rolf Theunissen <rolf.theunissen@gmail.com>