Initialize list fields with initial capacity 2
This micro optimization initializes the list fields with an initial size
of 2. When debugging the IDE with dark theme there it can be observed
that the used capacity of these fields is never larger than 2. The lists
would grow to an initial capacity of 10.
This will reduce a tiny bit of garbage during applying of CSS styles.

Change-Id: Iacac308026c2903029976511247e63d14e1ab11c
Signed-off-by: Karsten Thoms <karsten.thoms@karakun.com>