Bug 548430: [Cocoa] NullPointerException in Image.internal_new_GC
- There is an underlying assumption while creating breadcrumb arrows.
The image creation logic assumes the new image created is always created
with transparent color.

Currently only the opaque areas are set with alpha 255, which is opaque.
But the areas where transparency should be set is not handled at all.

This patch sets transparent area with alpha 0, which is transparent.

This change is based on https://git.eclipse.org/r/#/c/163181/

Change-Id: I371373a7d7e1825fb17032465ddbe4e31fb8c0db