Bug 563485 - Refactor StringMatcher into TextMatcher
Re-implement the matcher using org.eclipse.core.text.StringMatcher
and rename it to TextMatcher to avoid confusions. Fix the prefix
matching on individual words if wildcards are to be taken literally.

Add tests for the new TextMatcher.

The old StringMatcher is kept because it is used (despite being
internal!) in org.eclipse.equinox.p2.ui.discovery. Once it'll have
been replaced there by TextMatcher, StringMatcher can be removed.

Change-Id: Ib5bcb0617a6c894d0b0b2bff25d1d4ab279210a7
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>