Bug 404990 - [breakpoints] Delete breakpoint, undo, redo, undo. Gone.
The issue is located in both, the redo and the second undo operation.

Delete breakpoint and redo of this operation have the (for this issue
relevant) difference that the first set a 'breakpoint deleted' attribute
on the marker while the redo does not set/update this attribute.
The simple fix is to move the "mark deleted" into the delete marker
operation which is executed from an 'initial' delete as well as from a
redo.

After fixing that a breakpoint is restored but the view does not
updated.
BreakpointManager does not properly send added notifications. Most of
the time it works because an added breakpoint will be registered and the
registration will trigger the added notification.
However for the add of the second undo the registered attribute is
already set, the setRegister(true) which normally triggers the add
notification is not executed.

Change-Id: Ib509ec9c8116ac4a7bb3f8bcf38dc470e0b90b8b
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>