Bug 344352 - [breakpoints] Error dialog when deleting a working set
In fact there is no issue with deleting working set but a regular user
is most likely seeing this unnecessary error message from that action.
Deleting working set give the option to remove the working set, all
containing breakpoints or both. If removing breakpoints is unselected it
will still perform a remove operation for zero markers and the standard
remove-marker-operation will produce an error dialog because there are
no markers to delete.

Change-Id: If08c5f930bbdbf0236dd8a20af3f43b60b77d427
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>