Bug 563752 - [console] FileLink with offset need line number which is ignored
FileLink can be created with offset+length or line number to be selected
once the link is activated. All of those position parameters are
optional and there is no reason to specify offset+length and line number
on the same link.
However before this commit the offset+length was ignored if no line
number >= 0 was given.

Change-Id: Iaf105bc25acaec064763ebe60abf7cdcc409b58c
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>