Bug 477391 - Provide a deferred way to create ImageDescriptors in JFace
Since a majority of images are not used at any one time, calculating the
URL necessary for the image is an expensive (in terms of I/O) operation.
By using a Supplier<URL> instead, we can defer the lookup of the image's
URL until the image data is actually requested (which should be when the
image is displayed on screen).

Code that uses ImageDescriptor.createFromURL(expensiveURL) can be
refactored into ImageDescripor.createFromURLSupplier(()->expensiveURL)
to see performance gains.

Since this is new API, only callers that have been updated to use the
new method will see the benefit.

Change-Id: I3420c72dabf1d32234bed0b549fb70a246597b52
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>