Bug 475604 - DeclareImages in WorkbenchImages hits fs repeatedly
At startup, the workbench installs a number of images by scanning
bundles looking for appropriate images. This is triggered when the
bundle is started, and so hundreds of images are loaded before they are
needed during the splash screen window.

By moving the URL calculation until it is needed, when the image is
first accesed, we can defer the work until those specific icons are
shown on screen. This saves a fraction of a second of the startup time.

Change-Id: I3a77defc9c1a2e0ecab82fee4aa8c0c140aa24f6
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>