Bug 564153 - Fix EditorIconTest on macOS
The EditorIconTest bad icon image was showing a subtle variation between
the bad image and the known good one. This occurred because macOS is
loading the images on that platform, and the image data is reconstructed
on demand.

When we introduced the `DeferredImageDescriptor`, the implemenation used
the SWT parsing code for loading the images, and thus reflected the data
exactly as on disk. The macOS on the other hand was down-scaled from the
higher resolution image using `new Image(device, filenameprovider)` and
thus the slight blurs caused the digital signature to not match exactly.

Fix this by introducing one of the other overridden APIs in
`ImageDescriptor` so that the exact same codepath is used with the new
API as with the old one.

Change-Id: Idc8b5dc3854929279e16359e0830b9c4c40dd9cb
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>