Bug 564535 - Use Label widget for message text in dialog header
Change field type of messageLabel to Label. This removes the behavior
that the dialog's message is selectable, can have a cursor and context
menu from Text widgets.

The former consideration of a Text's inner trimmings is removed as a
Label takes the full client area.

Added a test snippet to test the dialog with a long wrapped message and
ability to test clipping behavior by resizing.

Change-Id: I98289d937673f7aab1aa5f202378ae3ec4b90559
Signed-off-by: Karsten Thoms <karsten.thoms@karakun.com>