Bug 563987 - adjust ranges and version for updated core.runtime package
Because the package org.eclipse.core.runtime is split between
org.eclipse.equinox.common and org.eclipse.equinox.registry and the
org.eclipse.core.runtime merges the contents into a single export with
all the parts together, the version of the the parts of the package need
to be kept in sync with each other.  Here org.eclipse.equinox.common
added new API which requires a bump the the package version and the
ranges for requiring the parts of the package.

Change-Id: Ic1997dd723ba354671e32776b8218dcae85d9aa7
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>