Bug 564062 - Refactor ContentTypeManager.DEBUGGING
The ContentTypeManager DEBUGGING field is initialised as soon as the class is
loaded. This may be too eager since it may not be needed until later.  By
moving the boolean into a static inner class, we can defer its initialisation
until the first time it is used in a code path.

We can also take advantage of Equinox's new `ServiceCaller` which allows for
single-shot access of services.

Change-Id: Iea9b52894f796c8b5cc6ad1fbdf248aafa745b4f
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>