Bug 558463 - [console] Console redirection changes content
Unfortunately all the existing streams monitoring/proxying/redirection
or basically most of the console IO stack is based upon strings.

If a program produce 'random' binary output and user configures console
to redirect the output to a file it might still be changed due to the
(until now) unavoidable string decoding.

Keeping existing API and not breaking clients made this change in some
parts more complex than it could be.

Change-Id: Id8e8fa1f777f06220af86dd4f026873c6b0b2d7b
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>