Bug 564305 - fixed unsorted filtered proposals on incremental complete
On incremental completion with async content assist and insert common
prefix enabled, the proposal order in fFilteredProposals did not match
the displayed order, causing unexpected proposals to be inserted.

The fComputedPropsals list will not be sorted in this scenario, so it
can't be used to overwrite fFilteredProposals.
fFilteredProposals seems to be always initialized by
computeAndPopulateProposals anyway.

Change-Id: I0da117281502b0a11ea462a38e129e3c59c252f1
Signed-off-by: Julian Honnen <julian.honnen@vector.com>