Bug 565674 - Cheat Sheets view doesn't open via Help > Cheat Sheets
Accessing cheat sheets via 'Help > Cheat Sheets...' does not open the
Cheat Sheets view. This is due to
CheatSheetCategoryBasedSelectionDialog.okPressed() hiding the shell of
the dialog and OpenCheatSheetAction using the active shell as parent, in
addition to the fix for bug 563555. Hiding the shell no longer results
in changing the active shell, and so the cheat sheet has the hidden
shell as a parent. As a result, nothing is shown to the user.

This change adds a new public method, with which the parent shell of
OpenCheatSheetAction can be specified. This is then used in
CheatSheetCategoryBasedSelectionDialog to pass the parent shell to
OpenCheatSheetAction.

Change-Id: I494dbb7d7865fd504d17a007617658f2472de192
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>