Bug 90757 - [Contributions] Outline view toolbar flashes on tab switch
When the toolbar is relayout in ToolBarManager, first the toolbar is
packed, then the parent is packed. The first pack paints the new toolbar
on the old location. The second pack moves the toolbar to the correct
location. This switching is observed as flicker. Disabling redrawing of
the parent while the pack calls are done removes the flicker.

Change-Id: I20ea9f392e402de2b236bf922ef43b2f6c8cb7f9
Signed-off-by: Rolf Theunissen <rolf.theunissen@gmail.com>