Bug 566375 - Disappearing 'View Menu' toolbar button in History view
When the view menu became available in the model, the topright was not
adjusted in StackRenderer. Fix to properly call the adjustTopRigth in
case children are added.

Furthermore, the view menu is not synchronized to the E4 model when
updateActionBars is called. Enabling this synchronization is too risky
now (RC1), therefor bypassing the model and renderers and forcing an UI
update of the topright.

Change-Id: If5fcf7f90d9257e3a48bae684746a23717b4b152
Signed-off-by: Rolf Theunissen <rolf.theunissen@gmail.com>