Bug 566375 - Disappearing 'View Menu' toolbar button in History view - Part 2
Forcing the UI update of the topright had has side effect that the
toolbar is not rendered at the correct location. When first updating the
toolbar and then forcing the UI update, the toolbar is correctly
renderer.
The forced UI update causes another flicker of the toolbar. This is
prevented by disabling redraw while repositioning the toolbar.

Change-Id: Iaceb41965f787f554139779300f7da11455c32f2
Signed-off-by: Rolf Theunissen <rolf.theunissen@gmail.com>