Bug 566336 - [Test] testAppendOutputToFile fails
Avoid a theoretical race condition where a fast terminating process get
not all content to console.

Includes a of-by-one error fix in MockProcess termination calculation.
Before a MockProcess with 0ms runtime was reported as not terminated for
up to 1ms. This can be relevant for the initial termination check in
RuntimeProcess.

This change might also fix random failures for other tests using the
doConsoleOutputTest method.

Change-Id: I76cad5317607b1889513827c8191edefa7637491
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>