Bug 565033 - Deadlock opening Ant build file
Changed SelectedResourceManager to avoid accessing UI thread via
syncExec() directly from the calling thread (that may own locks on which
UI thread is waiting already).

To avoid deadlocks, call to the UI thread is executed from a dedicated
thread that does not hold any locks with the max timeout of 10 seconds.
In case the UI thread doesn't react in this time, an error with stack
traces will be reported and the current call to SelectedResourceManager
will return null or empty selection.

Change-Id: I6f426900f71240afc4c0d5d51d692faea2aba1e3
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>