Bug 501718 - Modernized Eclipse Help UI prototype
This is a prototype that can be activated with the following Java
property:
-Dorg.eclipse.help.webapp.experimental.ui=m/index.html

For testing/debugging, further development, etc. see:
/org.eclipse.help.webapp/m/README.md

This prototype is based on HTML 5 with an iframe for the content and,
unlike the legacy UI, does not use HTML 4 framesets. Technically, it
sits on top of the legacy UI. For example, when doing a full search, the
HTML page of the legacy UI will be parsed and rendered in the web
browser using JavaScript. This has disadvantages, but it made it easier
to develop this prototype as a plugin without the need to patch Eclipse.
But once it is integrated into Eclipse, this should be changed (in
'm/README.md' see 'Further development').

Improvements compared to the legacy UI:
- Responsive, e.g. by automatically hiding the TOC bar on small screens
- Search as you type and auto completion of search terms
- Topic preview when hovering over instant search results
- Search scopes drop-down in the search field to search in all books, in
  the current book, in the current chapter or in a user-defined scope
- Full search results shown as a page instead of as side bar
- Full search results can be filtered by books
- Infocenter: UI designed aware of mobile devices
- Infocenter: URL contains topic or search, so a deep link can be
  bookmarked or shared in the usual way

The "Index" tab of the legacy UI was dropped in favor of a simpler UI.

Things to improve, where this prototype falls behind the legacy UI:
- Internationalization (by externalize Strings)
- Right-to-left (RTL) support
- Accessibility (HTML5 ARIA, keyboard support, color contrast, etc.)
- Dark theme support
- Customization, e.g. to be able to specify an Infocenter header/footer
- ...?

The prototype was developed as a plugin on GitHub and both authors
agreed to contribute it to Eclipse (see Signed-off-by below):
https://github.com/howlger/Eclipse-Help-Modernized

Change-Id: I8635deabf69a84ed293a9e89e2424b8be1212262
Signed-off-by: Harald Hermann <harald.hermann@advantest.com>
Signed-off-by: Holger Voormann <eclipse@voormann.de>