Bug 566549 - Avoid hard code dark colors in the dark theme and find better colors
This change should not result in any visual change in the default dark
theme but will allow the user to configure the two base colors of the
dark theme.

This is the first step in removing the hard-coded colors.

background-color:#515658;
color:#eeeeee;

Are now replaced by:

'#org-eclipse-ui-workbench-DARK_BACKGROUND',
'#org-eclipse-ui-workbench-DARK_FOREGROUND',

which can be styled by the user.


Change-Id: I3383d70fa76b9c30f54ae2bf8fea50d07b417d6c