Bug 493667 - [HiDPI] Workaround Internet Explorer font scaling bug
Problem:

The Microsoft Internet Explorer (and probably also legacy non Chromium
based Microsoft Edge versions) does scale system fonts (for instance
"font: icon") much too large on high-DPI displays, if no font size is
specified. This affects the Eclipse embedded help on Windows, since on
Windows the browser widget currently uses the Internet Explorer by
default.

In addition, "font: icon" without a font size affects also the
Infocenter since different browsers display it in different sizes. On
Windows 10, on my computer "font: icon;" without a font size is
displayed in Chrome and Edge in the font size 16px, on the same computer
it is displayed in IE and Firefox in 12px by default. Firefox and IE use
the system font used to label icons, but in Chrome changing the system
icon font does not change the font in the browser (note, in Windows 10
the UI to change the icon font and other fonts has been removed, so only
the general font size can be set without editing the registry). On the
iPad the font size seems to be 14px and on an Android smartphone it
seems to be 16px (Chrome) and 12px (Firefox).

The original idea to use "font: icon" without a font size to get as
close to the system look as possible does not seem to work anymore due
to the mentioned IE bug on high-DPI displays and the lack of support of
it in Chrome (including Microsoft Edge).

Solution:

As a compromise, add "0.875 rem" everywhere where "font: icon" and
"font: message-box" is used, which is 14px, as long as the default text
size of the web browser of 16px is not changed. This will change the
font size in most cases, but in each of these cases only a little bit
(either +2px or -2px).

Change-Id: Ic92118c37b2ba7eedcbc87e72e08d59cb727e525
Signed-off-by: Holger Voormann <eclipse@voormann.de>