Bug 567305 - Create a Java16 SE profile in Tycho
The issue is caused because JavaSE-1.8 doesn't contain org.w3c.dom.css
in standard; so this package is excluded from the list of visible
packages when compiling with --release 8. However, the
JavaSE-1.8.profile file (used by Tycho) does list this package as
visible, so it assumes the JRE provides it and doesn't add further
bundle to satisfy dependency to the package.


Change-Id: If326fbf124d803c48406dda709554f6a9808155d
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>