Bug 567898 - [JFace][HiDPI] ImageDescriptor support alternative naming scheme for high dpi
Currently ImageDescriptor automatically provides higher DPI images if a
resource exits that ends with @x2 or @1.5

This works well if only one resolution is used across the application
(e.g. 16x16 pixel icons) but rapidly blows up if multiple resolutions
are used as there are duplications. The second limitation is that only
two fixed zoom levels are supported at the moment.

The current approach can be enhanced to support another common scheme
where icon resources are organized in folders holding the resolution in
folder-name (e.g. icons/16x16/icon1.png, icons/32x32/icon1.png and so
on). This also has the advantage that it is possible to implement other
scaling factors beside the currently fixed ones without any code
changes.

This commit includes the following:

- add/extend test cases for both schemes (current + fallback) to ensure
compatibility with both schemes
- add proposed alternative naming scheme for high-dpi icons as a
fallback if the @x2/@x1.5 does not find a suitable resource


Change-Id: Ib156aedf9db8ef115ab8dd566731124f3c6e1d48
Signed-off-by: Christoph Läubrich <laeubi@laeubi-soft.de>