Bug 567974 - [Help] Option to show search results with their breadcrumbs
In "Search Results" tab add "Show Locations" button to toggle visibility
of the breadcrumbs.

Display breadcrumbs depending on the state of the "Show Locations"
button.

In the current modernized help UI prototype (see bug 501718), this
enables the following (the JavaScript code already exists in the current
prototype to parse and read the search results breadcrumbs from the
legacy UI in the format as implemented):
- Search results page: full expandable and selectable filter tree (full
tree, not only the books on the first level)
- Displaying of breadcrumbs (not optional; can not be disabled as in the
legacy UI)

Change-Id: I454a4f83f5a5b7128c854d195693d8ec1927fa4e
Signed-off-by: Holger Voormann <eclipse@voormann.de>
Signed-off-by: Wim Jongman <wim.jongman@remainsoftware.com>