Bug 378488 - "Open With > ..." should not store the default editor
"Open With > ..." currently stores the selected editor as default editor
for the selected file. This is wrong in 99.9% of usages according to the
Bug report to which I agree.

The main usage scenarios for Open With > ... are:
1.) Open the file with a different editor this time (but keep the
default).
2.) Always open similar files with the chosen editor.

Scenario 2 can be supported with bug 378485 if the "Open With >
Other..." dialog comes up with the current default editor pre-selected.

This simply removes this option without trying to add yet another power
user option via additional control key or additional menu items as we
already have too many of these hard to discover features.

Change-Id: Idb9a80309ffb408a3448a1f752a4ceb40aa0fdf9
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>