Bug 568204 - Replace Platform.getBundle with Platform.getLog
The caller in LaunchUpdateIntroAction uses Platform.getBundle in a cold
path in order to get hold of the platform log. However, this requires
looking up the bundle by name, when we already have a means to acquire
the log directly from the Platform, so use that instead.

Instead of using `log.(new Status)` use the default method `error` which
does the same thing and is more compact.

Change-Id: I822898f0ed91ed43a0390a4c0436024f93486bfd