Bug 568153 - Suggest alternate uses of Platform.getBundle
`Platform.getBundle(String)` provides a means to translate a String
bundle identifier into a `Bundle` instance. It performs a reasonable
amount of work, including iterating through all bundles, filtering and
other such processing.

It is generally more efficient, if you already have a class loaded by
OSGi, is to use `FrameworkUtil.getBundle(Class)` instead. This performs
a near direct lookup, since the class' associated `ClassLoader` implicitly
knows what the `Bundle` is.

However, the majority of cases is to write an entry to the log, from
which `Platform.getLog(Class)` is a much simpler implementation.

Add JavaDoc to introduce these patterns that may be of relevance to
callers instead.

Change-Id: I6a8bb35446660391eccec69061a98958d1a833ad
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>