Bug 568231 - Quick Search file pattern match doesn't work as expected
QuickSearch use the right text box to enter filename/path patterns to
limit the files to search in. It allows a comma separated list to
include multiple patterns.
Before this change it works only reasonable without spaces around the
separating commas. Possible spaces where not removed from the separated
pattern and a pattern with leading or trailing spaces rarely matches
anything. (that's why a single pattern is already trimmed)

Change-Id: I36ac86ede8ebec790c79a493dad517369f3ea72b
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>