Bug 568299 - rebuildRequested flag should cancel following build configs execution
If one config requested rebuild, running following configs in same round
only wastes resources. So we introduce a system flag that allows to skip
the rest of the current build round and start again with the first
config to be built.

The flag can be set via system property
-Dorg.eclipse.core.resources.allowEarlyInnerBuildLoopExit and by default
early exit is not allowed. This avoids potential side effects with
legacy software (so there will be no build behavior changes without
extra system property).

Change-Id: Ic59bf3227c98ee67bc937e7481d37d140d41c54e
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>