Bug 568328 - Change org.eclipse.core.filebuffers to use an util class instead of an activator
As other plug-ins may also access the string constant the name
FileBuffersPlugin is left as is.

The logging via the activator has been replaced by standard platform
logging. Also the access to the ITextFileBufferManager has been moved to
a static method and field, so the activator is not required anymore.

Change-Id: I394e121d7a65f82788231bb6b59287d6438cdb7a
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>