Bug 551260 - Clicking "Change Workspace" on "Older Workspace Version" dialog just redisplays the dialog
When running a new Eclipse version on an old workspace and if this
Eclipse installation is configured to use last or given workspace
without asking user to select one, the "Change Workspace" button from
upgrade workspace dialog does not work as expected.
Before this change it would just retry the same workspace again without
the ability to select another and of cause this would trigger again the
upgrade dialog.

Change-Id: Ida7fb47f26ce7beda2ab5bd689665c35ca95d8dc
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>