Bug 567618 - NPE in AdditionalInfoController.computeInformation
Either subclasses manage to change values on callbacks between we access
fields, or the code can be executed in parallel. I believe latter is the
case.

I didn't found a way to reproduce access from multiple threads, and do
not see that from code inspection, but the sporadicity of this issue is
most likely caused by MT access.

So changed affected fileds to volatile & reduced access to them to
single reads. No locks introduced, because without further information
how this bug could happen, this would be too dangerous.

Change-Id: Ifdaf2bea160eb314578529ccc6290b259dd18872
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>