Bug 568932 - [BigSur] Cannot hover over or inspect variables
On Big Sur, the empty inspect dialog pop-up is caused due to SWT Bug
567787- Setting Shell visible doesn't send SWT.SetData for Virtual
Tree/Table. Due to this the virtual tree in the pop-up is not populated
when set to visible.
The workaround is to call getText on the item so that it is already
created when pop-up is shown.

Change-Id: I241b7c1c75cbc600ad63b4e9d4bd491218d67619
Signed-off-by: Lakshmi Shanmugam <lshanmug@in.ibm.com>