Bug 569373: Use ViewFilter to filter launch configs
The content provider is only called when the launch configuration
tree is initialized. Any following add/remove of a launch
configuration is handled in LaunchConfigurationView without
fetching the new tree from the content provider.
Forcibly refresh the list in the viewer when the added launch
configuration is shared. To avoid having dangling launch
configurations in the tree.

Change-Id: Id67d550dddc6df583229013c1c71bcfc5f320f38
Signed-off-by: Torbjörn Svensson <azoff@svenskalinuxforeningen.se>